dev/carsense/carsense-main/carsense-main/backend/requirements.txt
fastapi>=0.103,<1.0
uvicorn[standard]>=0.23,<1.0
SQLAlchemy>=2.0,<3.0
alembic>=1.13,<2.0
pydantic>=2.6,<3.0
pydantic-settings>=2.2,<3.0
passlib[bcrypt]>=1.7,<2.0
python-jose[cryptography]>=3.3,<4.0
email-validator>=2.1,<3.0
python-multipart>=0.0.9,<1.0
requests>=2.31,<3.0
apscheduler>=3.10,<4.0
